requirements.txt
asgiref==3.7.2
Django==3.2.9
django-cors-headers==4.2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
phonenumbers==8.13.18
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
